requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pillow==11.0.0
pydantic==2.9.2
python-dotenv==1.0.1
bcrypt==4.2.1
gunicorn==23.0.0
